Repositiry Find bug fixed.

+++ Planner.Domain/Repositories/Interfaces/IRepository.cs
@@ -7,11 +7,11 @@ namespace Planner.Domain.Repositories.Interfaces
     public interface IRepository<T> where T : BaseEntity
     {
         IQueryable<T> GetAll();
-        Task<T> FindAsync(int id);
-        Task<T> GetByIdAsync(int id);
+        Task<T> FindAsync(long id);
+        Task<T> GetByIdAsync(long id);
         Task AddAsync(T entity);
         void Update(T entity);
         void Delete(T entity);
-        Task Delete(int id);
+        Task Delete(long id);
     }
 }
